Fix tests on Travis
Should fix #25.

The test was bad: it was assuming that (a) user 'root' exists, (b)
the test file is world-readable, and (c) every parent directory is
world-executable.  On Travis CI trusty containers one of these
assumptions appears to be incorrect.

It was also confusing the issue by using 'root', which is usually
special and can read anything, except that my file_readable_to() logic
doesn't implement that and treats it like any other user.